backendPython/requirements.txt
flask-sqlalchemy>=3.0.0pymysql
ariadne>=0.16.0
ariadne.contrib.federation>=0.16.0
flask>=2.0.0
flask-cors>=3.0.0